+++ WebApi.Contracts/Dto/AdvertDto.cs
@@ -14,7 +14,7 @@
         /// <summary>
         /// Стопимость / Price
         /// </summary>
-        public double Price { get; set; }
+        public uint Price { get; set; }
         /// <summary>
         /// Id cтатуca объявления / Id of a ads status
         /// </summary>
